+++ src/Rect.h
@@ -10,8 +10,8 @@
 class IntVector;
 class Rect {
 public:
-	int x, y, w, h;
-	Rect(int x, int y, int w, int h) : x(x), y(y), w(w), h(h) {
+	float x, y, w, h;
+	Rect(float x, float y, float w, float h) : x(x), y(y), w(w), h(h) {
 
 	}
 	Rect() : x(0), y(0), w(0), h(0) {
